Enable clang-diagnostic-sign-compare
305 sites over all 840 translation units, cleared with no suppressions. Roughly
240 of them fell to eight type changes rather than per-site casts, because the
recurring cause was not the comparisons -- it was that RA had already been
migrated to signed sizes and TD had not.

  TD DynamicVectorClass::Count() -> base::ssize        305 -> 216
  sizeof(x)/sizeof(x[0]) -> std::ssize(x), 26 sites    216 -> 190
  TD VectorClass interface -> base::ssize              190 -> 185
  TechnoTypeClass::Level and house counters -> int     185 -> 145
  radar geometry (RadarX/Y, RadarCell*, PixelPtr)      145 -> 101
  text-print coordinates -> int; retry/timeout -> long 101 -> 33
  Response_Time() family, FrameSendRate, BuildLevel,
  Scenario -> signed                                    33 -> 5
  starttime -> int64_t, mixfile bounds                   5 -> 0

RA's vector_dynamic.h already returned base::ssize from Count() while TD's
returned size_t; the same split existed for VectorClass, Level, and the radar
fields. That is why TD carried 170 of the 305 against RA's 103, and it is the
RA/TD divergence docs/CLANG_TIDY_PRIORITIES.md warns about.

td/vector.h's VectorMax deliberately stays unsigned. Migrating it with the rest
of the class widened it from 4 bytes to 8 and td_heap_layout_test failed on
three serialized types (56->64, 40->48, 1952->1960) -- the raw-byte save format.
The member keeps its width with a comment saying why; the public interface
around it is signed regardless, which is what the call sites compare against.
Full parity with RA there needs a SAVEGAME_VERSION bump and is left alone.

Several -1 sentinels were being compared against unsigned fields: Level != -1,
pos != -1, MaxRetries != -1. Those worked only by promoting -1 to UINT_MAX and
happening to match the sentinel's bit pattern. Most are now signed fields; the
two that stay bit patterns, in smudge.cc and 2keyfbuf.cc, say ~0U explicitly.

std::ssize rather than std::size for the array-length idiom: std::size is
unsigned and would leave the warning in place. One site could not use it at all
-- src/tech/sha.cc divides sizeof by sizeof over a union, not an array -- and
keeps an explicit cast. sha.h's block sizes moved from enumerators to
constexpr int, since sizeof() in an enumerator makes the whole unnamed enum
unsigned and every index comparison a mismatch.

The full-config sweep afterwards found ten second-order findings, all fixed
here: nine redundant casts that existed only because those fields had been
unsigned, and one "%03lu" that needed "%03td" once its argument became
base::ssize. It now reports 0 findings and no hard error: lines over all 840
units. Both games build, the 84 tests pass, and rasdl runs against the Steam
data.

+++ src/ra/display.cc
@@ -3553,7 +3553,8 @@ void DisplayClass::Mouse_Left_Release(CELL cell, int x, int y,
             **	formation offset, then it can't be a formation move.
             */
             const FootClass* foot = (FootClass*)tobject;
-            if (foot->Group != group || foot->XFormOffset == 0x80000000) {
+            if (foot->Group != group ||
+            foot->XFormOffset == static_cast<int>(0x80000000)) {
               FormMove = false;
               break;
             }
