+++ src/ra/edit.cc
@@ -164,8 +164,9 @@ void EditClass::Draw_Text(const char* text) {
   Conquer_Clip_Text_Print(text, X + 1, Y + 1, Color, TBLACK, TextFlags | flags,
                           Width - 2);
 
-  if (Has_Focus() && strlen(text) < MaxLength &&
-      String_Pixel_Width(text) + String_Pixel_Width("_") < Width - 2) {
+  if (Has_Focus() && static_cast<int>(strlen(text)) < MaxLength &&
+      static_cast<int>(String_Pixel_Width(text) + String_Pixel_Width("_")) <
+          Width - 2) {
     Conquer_Clip_Text_Print("_", X + 1 + String_Pixel_Width(text), Y + 1, Color,
                             TBLACK, TextFlags | flags);
   }
@@ -191,7 +192,8 @@ bool EditClass::Handle_Key(KeyASCIIType ascii) {
       break;
 
     default:
-      if (String_Pixel_Width(String) + Char_Pixel_Width(ascii) >= Width - 2) {
+      if (static_cast<int>(String_Pixel_Width(String) +
+                       Char_Pixel_Width(ascii)) >= Width - 2) {
         break;
       }
       if (Length >= MaxLength) {
