+++ src/ra/queue.cc
@@ -1354,7 +1354,7 @@ static RetcodeType Wait_For_Players(int first_time, ConnManClass* net,
  *   11/21/1995 BRR : Created.                                             *
  *=========================================================================*/
 static void Generate_Timing_Event(ConnManClass* net, int my_sent) {
-  unsigned long resp_time;  // connection response time, in ticks
+  long resp_time;  // connection response time, in ticks
   EventClass ev;
 
   //
@@ -1438,7 +1438,7 @@ static void Generate_Timing_Event(ConnManClass* net, int my_sent) {
  *   07/02/1996 BRR : Created.                                             *
  *=========================================================================*/
 static void Generate_Real_Timing_Event(ConnManClass* net, int my_sent) {
-  unsigned long resp_time;  // connection response time, in ticks
+  long resp_time;  // connection response time, in ticks
   EventClass ev;
   int highest_ticks;
   int i;
@@ -1558,7 +1558,7 @@ static void Generate_Real_Timing_Event(ConnManClass* net, int my_sent) {
 static void Generate_Process_Time_Event(ConnManClass* net) {
   EventClass ev;
   int avgticks;
-  unsigned long resp_time;  // connection response time, in ticks
+  long resp_time;  // connection response time, in ticks
 
   //
   // Measure the current connection response time.  This time will be in
@@ -3022,7 +3022,7 @@ int Extract_Uncompressed_Events(void* buf, int bufsize) {
   //------------------------------------------------------------------------
   // Loop until there are no more events in the packet
   //------------------------------------------------------------------------
-  while (leftover >= sizeof(EventClass)) {
+  while (leftover >= static_cast<int>(sizeof(EventClass))) {
     Keyboard->Check();
 
     event = (EventClass*)(static_cast<char*>(buf) + pos);
@@ -3115,7 +3115,8 @@ int Extract_Compressed_Events(void* buf, int bufsize) {
              sizeof(EventClass::EventType);
   event = (EventClass*)(static_cast<char*>(buf) + pos);
 
-  while (leftover >= datasize + sizeof(EventClass::EventType)) {
+  while (leftover >=
+         datasize + static_cast<int>(sizeof(EventClass::EventType))) {
     Keyboard->Check();
 
     //.....................................................................
