+++ src/ra/session.cc
@@ -1617,7 +1617,7 @@ unsigned long SessionClass::Compute_Unique_ID() {
   //------------------------------------------------------------------------
   path = getenv("PATH");
   if (path) {
-    for (i = 0; i < strlen(path); i++) {
+    for (i = 0; i < static_cast<int>(strlen(path)); i++) {
       Add_CRC(&id, static_cast<uint32_t>(path[i]));
     }
   }
