+++ src/td/edit.cc
@@ -350,8 +350,9 @@ void EditClass::Draw_Text(const char* text) {
     Conquer_Clip_Text_Print(text, X + 1, Y + 1, Color, TBLACK,
                             TextFlags | flags, Width - 2);
 
-    if (Has_Focus() && strlen(text) < MaxLength &&
-        String_Pixel_Width(text) + String_Pixel_Width("_") < Width - 2) {
+    if (Has_Focus() && static_cast<int>(strlen(text)) < MaxLength &&
+        static_cast<int>(String_Pixel_Width(text) + String_Pixel_Width("_")) <
+          Width - 2) {
       Conquer_Clip_Text_Print("_", X + 1 + String_Pixel_Width(text), Y + 1,
                               Color, TBLACK, TextFlags | flags);
     }
@@ -359,8 +360,9 @@ void EditClass::Draw_Text(const char* text) {
     Conquer_Clip_Text_Print(text, X + 1, Y + 1, Has_Focus() ? BLUE : WHITE,
                             TBLACK, TextFlags, Width - 2);
 
-    if (Has_Focus() && strlen(text) < MaxLength &&
-        String_Pixel_Width(text) + String_Pixel_Width("_") < Width - 2) {
+    if (Has_Focus() && static_cast<int>(strlen(text)) < MaxLength &&
+        static_cast<int>(String_Pixel_Width(text) + String_Pixel_Width("_")) <
+          Width - 2) {
       Conquer_Clip_Text_Print("_", X + 1 + String_Pixel_Width(text), Y + 1,
                               BLUE, TBLACK, TextFlags);
     }
@@ -427,7 +429,8 @@ bool EditClass::Handle_Key(KeyASCIIType ascii) {
       /*
       **	Don't add a character if the length is greater than edit width.
       */
-      if (String_Pixel_Width(String) + Char_Pixel_Width(ascii) >= Width - 2) {
+      if (static_cast<int>(String_Pixel_Width(String) +
+                       Char_Pixel_Width(ascii)) >= Width - 2) {
         break;
       }
