+++ src/tech/rndstraw.cc
@@ -174,7 +174,7 @@ void RandomStraw::Seed_Byte(char seed) {
  * HISTORY: * 07/10/1996 JLB : Created. *
  *=============================================================================================*/
 void RandomStraw::Seed_Short(short seed) {
-  for (int index = 0; index < sizeof(seed) * CHAR_BIT; index++) {
+  for (int index = 0; index < static_cast<int>(sizeof(seed) * CHAR_BIT); index++) {
     Seed_Bit(seed);
     seed >>= 1;
   }
@@ -194,7 +194,7 @@ void RandomStraw::Seed_Short(short seed) {
  * HISTORY: * 07/10/1996 JLB : Created. *
  *=============================================================================================*/
 void RandomStraw::Seed_Long(long seed) {
-  for (int index = 0; index < sizeof(seed) * CHAR_BIT; index++) {
+  for (int index = 0; index < static_cast<int>(sizeof(seed) * CHAR_BIT); index++) {
     Seed_Bit(seed);
     seed >>= 1;
   }
